requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dominate==2.7.0
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
fonttools==4.39.0
gitdb==4.0.10
GitPython==3.1.31
h5py==3.8.0
idna==3.4
imageio==2.26.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Keras-Applications==1.0.8
keras-ocr==0.9.2
kiwisolver==1.4.4
lazy_loader==0.1
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
pyclipper==1.3.0.post4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
tifffile==2023.2.28
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
validators==0.20.0
visitor==0.1.3
Werkzeug==2.2.3
zipp==3.15.0
